Rename shuttle-core crate to shuttle-runtime
