+++ shuttle-runtime/src/hint.rs
