+++ shuttle-runtime/src/runtime/task/labels.rs
@@ -66,7 +66,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// assert!(ext.insert(5i32).is_none());
     /// assert!(ext.insert(4u8).is_none());
@@ -84,7 +84,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// assert!(ext.get::<i32>().is_none());
     /// ext.insert(5i32);
@@ -103,7 +103,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// ext.insert(String::from("Hello"));
     /// ext.get_mut::<String>().unwrap().push_str(" World");
@@ -123,7 +123,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// *ext.get_or_insert(1i32) += 2;
     ///
@@ -139,7 +139,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// *ext.get_or_insert_with(|| 1i32) += 2;
     ///
@@ -160,7 +160,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// *ext.get_or_insert_default::<i32>() += 2;
     ///
@@ -177,7 +177,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// ext.insert(5i32);
     /// assert_eq!(ext.remove::<i32>(), Some(5i32));
@@ -195,7 +195,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// ext.insert(5i32);
     /// ext.clear();
@@ -214,7 +214,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// assert!(ext.is_empty());
     /// ext.insert(5i32);
@@ -230,7 +230,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext = Labels::new();
     /// assert_eq!(ext.len(), 0);
     /// ext.insert(5i32);
@@ -249,7 +249,7 @@ impl Labels {
     /// # Example
     ///
     /// ```
-    /// # use shuttle_core::current::Labels;
+    /// # use shuttle_runtime::current::Labels;
     /// let mut ext_a = Labels::new();
     /// ext_a.insert(8u8);
     /// ext_a.insert(16u16);
