+++ shuttle-schedulers/src/urw.rs
@@ -2,11 +2,11 @@ use rand::rngs::OsRng;
 use rand::seq::SliceRandom;
 use rand::{RngCore, SeedableRng};
 use rand_pcg::Pcg64Mcg;
-use shuttle_core::runtime::task::{Task, TaskId};
-use shuttle_core::scheduler::data::random::RandomDataSource;
-use shuttle_core::scheduler::data::DataSource;
-use shuttle_core::scheduler::{Schedule, Scheduler};
-use shuttle_core::seed_from_env;
+use shuttle_runtime::runtime::task::{Task, TaskId};
+use shuttle_runtime::scheduler::data::random::RandomDataSource;
+use shuttle_runtime::scheduler::data::DataSource;
+use shuttle_runtime::scheduler::{Schedule, Scheduler};
+use shuttle_runtime::seed_from_env;
 use std::collections::{HashMap, HashSet};
 use tracing::{trace, warn};
